--- a/test/abstractarray.jl
+++ b/test/abstractarray.jl
@@ -1258,3 +1258,13 @@ Base.pushfirst!(tpa::TestPushArray{T}, a::T) where T = pushfirst!(tpa.data, a)
     pushfirst!(tpa, 6, 5, 4, 3, 2)
     @test tpa.data == reverse(collect(1:6))
 end
+
+@testset "copyto! with tuple" begin
+    randtype(n) = rand(Bool) ? 1.0 : 2
+    @test copyto!(fill(0.0, 100), ntuple(randtype, 100))[end] != 0.0
+    @test copyto!(fill(0.0, 100), ntuple(x->1.0, 100))[end] != 0.0
+    @test copyto!(fill(0.0, 100), ntuple(randtype, 50))[end] == 0.0
+    @test_throws BoundsError copyto!(fill(0.0, 50), ntuple(randtype, 100))
+    @test_throws BoundsError copyto!(fill(0.0, 50), ntuple(x->1.0, 100))
+    @test_throws ArgumentError copyto!(fill(0.0, 5), ntuple(randtype, 7))
+end