@Keno 2020-12-30
Would make a good nanosoldier benchmark if somebody wanted to add it over there