@simon-dedeo 2020-12-30
Hey, I'm that guy. Something might already have been fixed here, since people are reporting that the delays are not as bad with Julia 1.6 (I found the problem just by accident, using Julia 1.5). See, for example, here: https://twitter.com/LapeyreJohn/status/1344123040762896384?s=20