@Keno 2020-12-30
The comparison in that tweet is with `Any[1, 2.0]` which was always fast (and more closely matches the semantics of Ruby, which doesn't have typed arrays). Without the `Any`, the element type needs to be computed from the literals, which is more complicated and is the case spend up by this PR.