@DilumAluthge 2020-12-30
> Would make a good nanosoldier benchmark if somebody wanted to add it over there

Who's the right person to ping for adding a benchmark to Nanosoldier? @christopher-dG, maybe?