@Keno 2020-12-30
It's easy to do, they just need to be added https://github.com/JuliaCI/BaseBenchmarks.jl - I just wanted to give somebody new the opportunity to learn how to do that ;)