@vtjnash 2021-01-05
The new test is sometimes failing on CI (https://build.julialang.org/#/builders/34/builds/7719/steps/5/logs/stdio) with probability of precisely 1/64 (the probability of ntuple being homogeneous).

Closer inspection seems to suggest the new code might be a poor model for the expected behavior of `copyto!`? This is also apparent in the other two tests, which are throwing BoundsError now, while previously they were expected to throw an ArgumentError. This is because the PR assumes that all AbstractArrays are IndexLinear and changes the assignment to the destination slice to be computed as `dst[begin:1:n] .= src`. Is that valid? Previously, the implementation would use a loop with eachindex to copy the elements. The test failure happens because the `@generated` path is emulating the old error behavior that falls out of the old design, while the new behavior (and especially the `@generated` fallback path) are now defined to encounter slightly different behaviors and failures.